LLMRouter/backend/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-mistralai
langsmith
qdrant-client
python-multipart
pydantic
python-dotenv
bcrypt==4.0.1
PyJWT
langchain-qdrant
pypdf
google-generativeai
mistralai